admin_bot/requirements/requirements.txt
discord.py
opencv-python
psycopg[binary]
requests
PyYAML
pillow
pytz==2022.7
ics==0.8.0.dev0
